app/requirements.txt
flask
flask-socketio
openai
gunicorn
python-dotenv